pyproject.toml
[build-system]
requires      = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name            = "limes_outpost"
version         = "0.1.0"
description     = "LimesOutpost — Autonomous Content OS"
requires-python = ">=3.13"
dependencies = [
    "jsonschema>=4.0.0",
    "python-dotenv",
    "psycopg2-binary",
    "openai",
    "pydantic",
    "requests",
    "newsdataapi",
    "PyJWT",
    "google-api-python-client>=2.0.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.0.0",
    "requests-oauthlib>=1.3.0",
    "pydantic-settings>=2.0.0",
    # Phase 3 additions (add when ready):
    # "celery[redis]>=5.3",
    # "redis>=5.0",
    # "flower>=2.0",
    # Phase 4 additions:
    # "fastapi>=0.110",
    # "uvicorn[standard]>=0.29",
    # "python-jose[cryptography]>=3.3",
    # "passlib[bcrypt]>=1.7",
    # "httpx>=0.27",
]

[project.scripts]
limes_outpost = "cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["limes_outpost*", "cli*", "worker*", "api*"]

[tool.setuptools.package-data]
"limes_outpost.contracts" = ["*.json"]
"limes_outpost"           = ["py.typed"]
